PrepararAmbiente/requirements.txt
mysql-connector-python==8.0.29
openpyxl==3.0.10
pandas==1.4.2
pip-chill==1.0.1
pymysql==1.0.2
